First test.

+++ src/platform.h
@@ -1,10 +1,11 @@
 #ifndef _PLATFORM_H
 #define _PLATFORM_H
+#include <stdint.h>
 
 #define TICKRATE 20000000
 
-typedef unsigned long timeval_t;
-typedef unsigned int degrees_t;
+typedef uint32_t timeval_t;
+typedef uint16_t degrees_t;
 /* timeval_t is gauranteed to be 32 bits */
 
 timeval_t current_time();
